% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_html_pack.R
\name{create_html_pack}
\alias{create_html_pack}
\title{create_html_pack}
\usage{
create_html_pack(
  output_list,
  filepath = "./created_html_pack.html",
  title = "Title",
  clean_up = TRUE
)
}
\arguments{
\item{output_list}{named list - Contains the output to export to html.
The list can be nested however all elements of all sub-lists must also be naemd}

\item{filepath}{string - file path and name of html. Best practive should end in html}

\item{title}{string - title to be used for the pack. Default is no title}

\item{clean_up}{logical - should intermediary files (.Rmd and .rds) be removed. Default is \code{TRUE}.
Could be changed to \code{FALSE} for debugging to to manually edit the pack e.g. adding comments.}
}
\value{

}
\description{
Create an html pack from a (possibly nested) named list(s).
The list(s) can contain plots, dataframes or other R objects.
Each list must be named as the names of each element is used as section headers.
The html is created via writing an rmarkdown file which is then knitted and deleted (assuming \code{clean_up=TRUE})
}
\examples{

\donttest{
set.seed(666)

output_list <- list()
output_list[["table 1"]] <- data.frame(a=runif(25), b=runif(25))
output_list[["Plot 1"]] <- ggplot(data=output_list[["table 1"]]) + geom_point(aes(x=~a, y=~b))

create_html_pack(output_list = output_list, filepath="./create_html_pack.html")
}

}
