% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expl_categorical.R
\name{expl_categorical}
\alias{expl_categorical}
\title{expl_categorical}
\usage{
expl_categorical(df, char.level = 20, num.level = 20)
}
\arguments{
\item{df}{dataframe - train set.}

\item{char.level}{Numeric - exclude character|factor columns if levels > char.level}

\item{num.level}{Numeric - exclude numeric|integer columns if levels > num.level (NULL discards numeric columns)}
}
\value{
a dataframe containing the columns:
var [Character]: a character variable
level [Character]: a level for that character variable
class [Character]: what class is the variable
n [Integer]: count of that level
perc [Numeric]: percentage of that level

no character columns: "no character or factors"
too many levels in all character columns: "no columns with fewer levels than char.level"
}
\description{
When given a dataframe object, this function will return
a dataframe object with 1 row per level for variable
sorted alphabetically.

It will provide:
how many rows contain that level
and the percentage of that value for that variable.
}
\examples{
df <- data.frame(pet = rep(c('dog', 'cat', 'horse', 'hamster'), c(40,30,20,10)),
                age = 1:100,
                size = factor(rep(c('Small','Large', NA), c(50,30,20)), levels = c('Small', 'Large'), ordered = TRUE),
                stringsAsFactors = FALSE)
expl_categorical(df, char.level = 3)
expl_categorical(df, char.level = 5)

}
