% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expl_summary.R
\name{expl_summary}
\alias{expl_summary}
\title{expl_summary}
\usage{
expl_summary(
  .data,
  summarise_vars = NULL,
  group_vars = NULL,
  remove_missing = TRUE,
  row_limit = 10000,
  order_col_by_stat = FALSE
)
}
\arguments{
\item{.data}{Dataframe or data table}

\item{summarise_vars}{Vector of numeric or integer column name strings to summarize.}

\item{group_vars}{Vector of column name strings to group by.}

\item{remove_missing}{Boolean for na.rm in mean and standard deviation.}

\item{row_limit}{Integer row limit.}

\item{order_col_by_stat}{Boolean to sort output columns sorted by statistic. Useful for comparing similar variables in summarise_vars.}
}
\value{
a dataframe containing the columns:
One column per grouped variable: a character variable,
mean [Numeric]: The mean of the summarized variable,
sd [Numeric]: The standard deviation of the summarized variable,
n [Integer]: The count of the summarized variable,
na_count [Integer]: The count of NAs in the summarized variable.
}
\description{
Will give the mean, standard deviation, length, and NA count for the summarized variable.
Variables can be added to create groupings.
}
\examples{
set.seed(123)
df <- data.frame(score_raw = rnorm(mean = 10, n = 100),
                 dataset = factor(rep(c("VAL", "TRAIN", "TEST"), c(20, 70, 10)),
                                  levels = c("TRAIN", "VAL", "TEST")),
                 target = rep(c(1, 0), c(80, 20)),
                 stringsAsFactors = FALSE)

df$score10 <- cut(df$score_raw,
                  breaks = quantile(df$score_raw, seq(0, 1, 0.1)),
                  labels = 1:10,
                  include.lowest = TRUE)

df \%>\% expl_summary(summarise_vars = "score_raw")
df \%>\% expl_summary(summarise_vars = "score_raw", group_vars = "dataset")
df \%>\% expl_summary(summarise_vars = c("score_raw", "target"), group_vars = c("dataset", "score10"))

}
