% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_confusion_matrix.R
\name{metric_confusion_matrix}
\alias{metric_confusion_matrix}
\title{metric_confusion_matrix}
\usage{
metric_confusion_matrix(
  actual,
  predicted,
  weight = rep(1, length(actual)),
  na.rm = FALSE,
  threshold = 0.5
)
}
\arguments{
\item{actual}{Array[Numeric] - Values we are aiming to predict.}

\item{predicted}{Array[Numeric] - Values that we have predicted.}

\item{weight}{Optional: Array[Numeric] - Weighting of predictions. If NULL even weighting is used}

\item{na.rm}{Optional: boolean - If \code{FALSE} function will return NA is any value in NA}

\item{threshold}{Optional: Numeric between 0 and 1. If prediction proablity is below \code{threshold} the predicted value is 0.}
}
\value{
confusion matrix for the classification. Col names are \code{c("actual.1", "actual.0")} and Row names are c("predicted.1", "predicted.0")
}
\description{
Returns a confusion matrix showing true(/false) positives(/negatives)
Note: Predictions should be annualized (independent of exposure)
}
\section{Inputs}{

}

\examples{

metric_confusion_matrix(actual=c(0,1,0,0), predicted=c(0.1,0.9,0.4,0.6))

}
