% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_deviance.R
\name{metric_deviance}
\alias{metric_deviance}
\title{metric_deviance}
\usage{
metric_deviance(
  actual,
  predicted,
  weight = rep(1, length(actual)),
  family = "gaussian",
  na.rm = FALSE,
  rebase = FALSE,
  tweedie_power = NULL
)
}
\arguments{
\item{actual}{Array[Numeric] - Values we are aiming to predict.}

\item{predicted}{Array[Numeric] - Values that we have predicted.}

\item{weight}{Optional: Array[Numeric] - Weighting of predictions. If NULL even weighting is used}

\item{family}{String - family of distribution. Must be in  \code{c("gaussian", "poisson", "gamma", "tweedie", "binomial")}}

\item{na.rm}{Optional: boolean - If \code{FALSE} function will return NA is any value in NA}

\item{rebase}{Optional: boolean - If \code{TRUE} predictions weighted mean will be rebased to match actual mean}

\item{tweedie_power}{Numeric - power of tweedie distribution if chosen in \code{family}}
}
\value{
Numeric: deviance of the predictions
}
\description{
Returns The deviance of the predictions.
Note: Predictions should be annualized (independent of exposure)
Note: Low is good. 0 is perfect model
}
\section{Inputs}{

}

\examples{

set.seed(666)
actual <- rep(10, 10)
predicted <- rnorm(n = 10, mean = 10, sd = 1)
weight <- pmax(rnorm(n = 10, mean = 10, sd = 1) , 0)

metric_deviance(actual, predicted, family="gaussian")
metric_deviance(actual, predicted, weight, family="gaussian")

}
\seealso{
Other Metrics: 
\code{\link{metric_mae}()},
\code{\link{metric_nloglik}()},
\code{\link{metric_pode}()},
\code{\link{metric_pove}()},
\code{\link{metric_rmse}()}
}
\concept{Metrics}
