% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_error_checking.R
\name{metric_error_checking_family}
\alias{metric_error_checking_family}
\title{metric_error_checking_family}
\usage{
metric_error_checking_family(
  actual,
  predicted,
  weight = rep(1, length(actual)),
  family = "gaussian",
  na.rm = FALSE,
  rebase = FALSE,
  tweedie_power = NULL
)
}
\arguments{
\item{actual}{Array[Numeric] - Values we are aiming to predict.}

\item{predicted}{Array[Numeric] - Values that we have predicted.}

\item{weight}{Optional: Array[Numeric] - Weighting of predictions. If NULL even weighting is used}

\item{family}{String - family of distribution. Must be in  \code{c("gaussian", "poisson", "gamma", "tweedie", "binomial")}}

\item{na.rm}{Optional: boolean - If \code{FALSE} function will return NA is any value in NA}

\item{rebase}{Optional: boolean - If \code{TRUE} predictions weighted mean will be rebased to match actual mean}

\item{tweedie_power}{Numeric - power of tweedie distribution if chosen in \code{family}}
}
\description{
function containing all of the error catching for calculating model metrics
This includes vectors being of the correct length and type, actuals and predictions agreeing with the distribution family
and the distribution family being allowed
}
\section{Inputs}{

}

