% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_error_checking.R
\name{metric_error_checking_nofamily}
\alias{metric_error_checking_nofamily}
\title{metric_error_checking_nofamily}
\usage{
metric_error_checking_nofamily(
  actual,
  predicted,
  weight = rep(1, length(actual)),
  na.rm = FALSE,
  rebase = FALSE
)
}
\arguments{
\item{actual}{Array[Numeric] - Values we are aiming to predict.}

\item{predicted}{Array[Numeric] - Values that we have predicted.}

\item{weight}{Optional: Array[Numeric] - Weighting of predictions. If NULL even weighting is used}

\item{na.rm}{Optional: boolean - If \code{FALSE} function will return NA is any value in NA}

\item{rebase}{Optional: boolean - If \code{TRUE} predictions weighted mean will be rebased to match actual mean}
}
\description{
function containing all of the error catching for calculating model metrics
This includes vectors being of the correct length and type
}
\section{Inputs}{

}

