% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_Fbeta.R
\name{metric_Fbeta}
\alias{metric_Fbeta}
\title{metric_Fbeta}
\usage{
metric_Fbeta(
  actual,
  predicted,
  weight = rep(1, length(actual)),
  na.rm = FALSE,
  threshold = 0.5,
  beta = 1
)
}
\arguments{
\item{actual}{Array[Numeric] - Values we are aiming to predict.}

\item{predicted}{Array[Numeric] - Values that we have predicted.}

\item{weight}{Optional: Array[Numeric] - Weighting of predictions. If NULL even weighting is used}

\item{na.rm}{Optional: boolean - If \code{FALSE} function will return NA is any value in NA}

\item{threshold}{Optional: Numeric between 0 and 1. If prediction proablity is below \code{threshold} the predicted value is 0.}

\item{beta}{Numeric}
}
\value{
precision of classification TP / (TP + FN)
}
\description{
Returns the Fbeta [(1 + beta^2) * (precision * recall) / ((beta^2 *precision) + recall)] of a classification using the confusion matrix
Note: Predictions should be annualized (independent of exposure)
Note: Perfect Fbeta is 1, poor model is 0
}
\section{Inputs}{

}

\examples{

metric_Fbeta(actual=c(0,1,0,0), predicted=c(0.1,0.9,0.4,0.6))
metric_Fbeta(actual=c(0,1,0,0), predicted=c(0.1,0.9,0.4,0.6), threshold=0.7)

}
\seealso{
\code{\link{metric_precision}}, \code{\link{metric_recall}} and \code{\link{metric_confusion_matrix}}
}
