% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_gini.R
\name{metric_gini}
\alias{metric_gini}
\title{metric_gini}
\usage{
metric_gini(actual, predicted, weight = rep(1, length(actual)), na.rm = FALSE)
}
\arguments{
\item{actual}{Array[Numeric] - Values we are aiming to predict.}

\item{predicted}{Array[Numeric] / DataFrame[Numeric] - Values that we have predicted.}

\item{weight}{Optional: Array[Numeric] - Weighting of predictions. If NULL even weighting is used.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
gini coeffient. Single value if \code{predicted} is vector. Named list if \code{predicted} is dataframe.
}
\description{
Returns gini coeffient of one or more models

Note: Predictions should be annualised (independent of exposure)
}
\examples{

# Input as vector
actual <- rnorm(100, mean=100, sd=10)
weight <- rep(1,100)
predicted <- actual + rnorm(100, mean=0, sd=1)

metric_gini(actual, predicted, weight)

# Input as dataframe

predicted <- data.frame(pred1 = actual + rnorm(100, mean=0, sd=1), pred2 = rnorm(100, mean=0, sd=1))
metric_gini(actual, predicted, weight)

}
\seealso{
\code{\link{plot_gini}}
}
