% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_mae.R
\name{metric_mae}
\alias{metric_mae}
\title{metric_mae}
\usage{
metric_mae(
  actual,
  predicted,
  weight = rep(1, length(actual)),
  na.rm = FALSE,
  rebase = FALSE
)
}
\arguments{
\item{actual}{Array[Numeric] - Values we are aiming to predict.}

\item{predicted}{Array[Numeric] - Values that we have predicted.}

\item{weight}{Optional: Array[Numeric] - Weighting of predictions. If NULL even weighting is used}

\item{na.rm}{Optional: boolean - If \code{FALSE} function will return NA is any value in NA}

\item{rebase}{Optional: boolean - If \code{TRUE} predictions weighted mean will be rebased to match actual mean}

\item{family}{can be NULL has this metric doesn't depend on family}
}
\value{
Numeric: value of mean absolute error
}
\description{
Returns the mean absolute error for a set of predictions.
Note: Predictions should be annualized (independent of exposure)
Note: Low is good
}
\section{Inputs}{

}

\examples{

set.seed(666)
actual <- rep(10, 10)
predicted <- rnorm(n = 10, mean = 10, sd = 1)
weight <- pmax(rnorm(n = 10, mean = 10, sd = 1) , 0)

metric_mae(actual, predicted)
metric_mae(actual, predicted, weight)

}
\seealso{
Other Metrics: 
\code{\link{metric_deviance}()},
\code{\link{metric_nloglik}()},
\code{\link{metric_pode}()},
\code{\link{metric_pove}()},
\code{\link{metric_rmse}()}
}
\concept{Metrics}
