% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_nloglik.R
\name{metric_nloglik}
\alias{metric_nloglik}
\title{metric_nloglik}
\usage{
metric_nloglik(
  actual,
  predicted,
  weight = rep(1, length(actual)),
  family = "gaussian",
  na.rm = FALSE,
  rebase = FALSE,
  tweedie_power = NULL
)
}
\arguments{
\item{actual}{Array[Numeric] - Values we are aiming to predict.}

\item{predicted}{Array[Numeric] - Values that we have predicted.}

\item{weight}{Optional: Array[Numeric] - Weighting of predictions. If NULL even weighting is used}

\item{family}{String - family of distribution. Must be in  \code{c("gaussian", "poisson", "gamma", "tweedie", "binomial")}}

\item{na.rm}{Optional: boolean - If \code{FALSE} function will return NA is any value in NA}

\item{rebase}{Optional: boolean - If \code{TRUE} predictions weighted mean will be rebased to match actual mean}

\item{tweedie_power}{Numeric - power of tweedie distribution if chosen in \code{family}}
}
\value{
Numeric: negative log likelihood
}
\description{
Returns the negative log likelihood for the predictions.
Note: Predictions should be annualized (independent of exposure)
Note: Large negative values are good. (function should be minimised)
}
\section{Inputs}{

}

\examples{

set.seed(666)
actual <- pmax(rep(10, 10), 0)
predicted <- pmax(rnorm(n = 10, mean = 10, sd = 1), 0)
weight <- pmax(rnorm(n = 10, mean = 10, sd = 1), 0)

metric_nloglik(actual, predicted, family="poisson")
metric_nloglik(actual, predicted, weight, family="poisson")
metric_nloglik(actual, predicted, weight, family="gamma")
metric_nloglik(actual=c(rep(0,5), rep(1,5)), predicted=seq(0.01,0.99,length.out=10), weight, family="binomial")

}
\seealso{
Other Metrics: 
\code{\link{metric_deviance}()},
\code{\link{metric_mae}()},
\code{\link{metric_pode}()},
\code{\link{metric_pove}()},
\code{\link{metric_rmse}()}
}
\concept{Metrics}
