% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_pode.R
\name{metric_pode}
\alias{metric_pode}
\title{metric_pode}
\usage{
metric_pode(
  actual,
  predicted,
  weight = rep(1, length(actual)),
  family = "gaussian",
  na.rm = FALSE,
  rebase = FALSE,
  tweedie_power = NULL
)
}
\arguments{
\item{actual}{Array[Numeric] - Values we are aiming to predict.}

\item{predicted}{Array[Numeric] - Values that we have predicted.}

\item{weight}{Optional: Array[Numeric] - Weighting of predictions. If NULL even weighting is used}

\item{family}{String - family of distribution. Must be in  \code{c("gaussian", "poisson", "gamma", "tweedie", "binomial")}}

\item{na.rm}{Optional: boolean - If \code{FALSE} function will return NA is any value in NA}

\item{rebase}{Optional: boolean - If \code{TRUE} predictions weighted mean will be rebased to match actual mean}

\item{tweedie_power}{Numeric - power of tweedie distribution if chosen in \code{family}}
}
\value{
Numeric: value of proportion of deviance
}
\description{
Returns the proportion of deviance explained by the predictions.
Note: Predictions should be annualized (independent of exposure)
Note: high is good. 1 is perfect model, 0 is null model, negative is worse than null model
}
\section{Inputs}{

}

\examples{

set.seed(666)
actual <- rnorm(n = 10, mean = 10, sd = 3)
predicted <- actual + rnorm(n = 10, mean = 0, sd = 1)
weight <- pmax(rnorm(n = 10, mean = 10, sd = 1) , 0)

metric_pode(actual, predicted)
metric_pode(actual, predicted, weight)
metric_pode(actual, predicted, weight, family="gamma")

}
\seealso{
Other Metrics: 
\code{\link{metric_deviance}()},
\code{\link{metric_mae}()},
\code{\link{metric_nloglik}()},
\code{\link{metric_pove}()},
\code{\link{metric_rmse}()}
}
\concept{Metrics}
