% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_pove.R
\name{metric_pove}
\alias{metric_pove}
\title{metric_pove}
\usage{
metric_pove(
  actual,
  predicted,
  weight = rep(1, length(actual)),
  na.rm = FALSE,
  rebase = FALSE
)
}
\arguments{
\item{actual}{Array[Numeric] - Values we are aiming to predict.}

\item{predicted}{Array[Numeric] - Values that we have predicted.}

\item{weight}{Optional: Array[Numeric] - Weighting of predictions. If NULL even weighting is used}

\item{na.rm}{Optional: boolean - If \code{FALSE} function will return NA is any value in NA}

\item{rebase}{Optional: boolean - If \code{TRUE} predictions weighted mean will be rebased to match actual mean}
}
\value{
Numeric: value of proportion of variance explained
}
\description{
Returns the proportion of variance explained by the predictions.
Note: Predictions should be annualized (independent of exposure)
Note: high is good. 1 is perfect model, 0 is null model, negative is worse than null model
}
\section{Inputs}{

}

\examples{

set.seed(666)
actual <- rnorm(n = 10, mean = 10, sd = 3)
predicted <- actual + rnorm(n = 10, mean = 0, sd = 1)
weight <- pmax(rnorm(n = 10, mean = 10, sd = 1) , 0)

metric_pove(actual, predicted)
metric_pove(actual, predicted, weight)

}
\seealso{
Other Metrics: 
\code{\link{metric_deviance}()},
\code{\link{metric_mae}()},
\code{\link{metric_nloglik}()},
\code{\link{metric_pode}()},
\code{\link{metric_rmse}()}
}
\concept{Metrics}
