% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_ROC.R
\name{metric_ROC}
\alias{metric_ROC}
\title{metric_ROC}
\usage{
metric_ROC(actual, predicted, weight = rep(1, length(actual)), na.rm = FALSE)
}
\arguments{
\item{actual}{Array[Numeric] - 0 or 1 - Values we are aiming to predict.}

\item{predicted}{Array[Numeric] / DataFrame[Numeric] - Between 0 and 1 - Values that we have predicted.}

\item{weight}{Optional: Array[Numeric] - Weighting of predictions. If NULL even weighting is used.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
Area under the ROC curve. Single value if \code{predicted} is vector. Named list if \code{predicted} is dataframe.
}
\description{
Return the area under the ROC curve
}
\examples{

data <- data.frame(x1=runif(100), x2=runif(100), noise=rnorm(100, sd=0.2)) \%>\%
  mutate(target=ifelse(x1 + noise>0.5, 1, 0))

metric_ROC(actual=data$target, predicted=data$x1)
metric_ROC(actual=data$target, predicted=data[c("x1","x2")])
}
\seealso{
\code{\link{plot_ROC}}
}
