% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_gini.R
\name{null_gini}
\alias{null_gini}
\title{null_gini}
\usage{
null_gini(actual, predicted, weight = rep(1, length(actual)), na.rm = FALSE)
}
\arguments{
\item{actual}{Array[Numeric] - Values we are aiming to predict.}

\item{predicted}{Array[Numeric] - Values that we have predicted.}

\item{weight}{Optional: Array[Numeric] - Weighting of predictions. If NUll even weighting is used}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
dataframe with columns
prediction - value of predictions (sorted)
cum_act - cumulative fraction of actuals
cum_pop - cumulative fraction of weight
}
\description{
Return the curve for a gini curve for threshold in the prediction. Need to plot gini curve

Note: Predictions should be annualised (independent of exposure)
}
\examples{

actual <- rnorm(100, mean=100, sd=10)
weight <- rep(1,100)
predicted <- actual + rnorm(100, mean=0, sd=1)

null_gini(actual, predicted, weight)

}
