% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_PrecisionRecall.R
\name{null_PrecisionRecall}
\alias{null_PrecisionRecall}
\title{null_PrecisionRecall}
\usage{
null_PrecisionRecall(
  actual,
  predicted,
  weight = rep(1, length(actual)),
  na.rm = FALSE
)
}
\arguments{
\item{actual}{Array[Numeric] - 0 or 1 - Values we are aiming to predict.}

\item{predicted}{Array[Numeric] - Between 0 and 1 - Values that we have predicted.}

\item{weight}{Optional: Array[Numeric] - Weighting of predictions. If NULL even weighting is used.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
dataframe with columns
Threshold - value of the threshold
Precision - Precision at each threshold
Recall - Recall at each threshold
}
\description{
Return the precision and recall for each threshold in the prediction. Need to plot Precision-Recall curve
}
\examples{

data <- data.frame(x1=runif(100), x2=runif(100), noise=rnorm(100, sd=0.2)) \%>\%
  mutate(target=ifelse(x1 + noise>0.5, 1, 0))

null_PrecisionRecall(actual=data$target, predicted=data$x1)
}
\seealso{
\code{\link{plot_PrecisionRecall}} and \code{\link{metric_PrecisionRecall}}
}
