% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_ROC.R
\name{null_ROC}
\alias{null_ROC}
\title{null_ROC}
\usage{
null_ROC(actual, predicted, weight = rep(1, length(actual)), na.rm = FALSE)
}
\arguments{
\item{actual}{Array[Numeric] - 0 or 1 - Values we are aiming to predict.}

\item{predicted}{Array[Numeric] / DataFrame[Numeric] - Between 0 and 1 - Values that we have predicted.}

\item{weight}{Optional: Array[Numeric] - Weighting of predictions. If NULL even weighting is used}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
dataframe with columns
Threshold - value of the threshold
TP_Rate - True positive rate
FP_Rate - False positive rate
}
\description{
Return the true positive rate and false positive rate for each threshold in the prediction. Need to plot ROC curve
}
\examples{

data <- data.frame(x1=runif(100), x2=runif(100), noise=rnorm(100, sd=0.2)) \%>\%
  mutate(target=ifelse(x1 + noise>0.5, 1, 0))

null_ROC(actual=data$target, predicted=data$x1)
}
\seealso{
\code{\link{plot_ROC}} and \code{\link{metric_ROC}}
}
