% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2way_comparison.R
\name{plot_2way_comparison}
\alias{plot_2way_comparison}
\title{plot_2way_comparison}
\usage{
plot_2way_comparison(
  actual,
  incumbent_pred,
  proposed_pred,
  weight = rep(1, length(actual)),
  n_bins = 9,
  incumbent_label = "Incumbent",
  proposed_label = "Proposed",
  min_weight = 0,
  rebase = FALSE
)
}
\arguments{
\item{actual}{Array[Numeric] - Values we are aiming to predict.}

\item{incumbent_pred}{Array[Numeric] - Values that we have predicted by the incumbent model. Incumbent is the current model looking to be replaced.}

\item{proposed_pred}{Optional: String - Text to use to label the proposed predictions.}

\item{weight}{Optional: Array[Numeric] - Weighting of predictions. If NULL even weighting is used.}

\item{n_bins}{integerish: number bins to split data into. Works best if n_bins is square number}

\item{incumbent_label}{Optional: String - Text to use to label the incumbent predictions.}

\item{min_weight}{Optional: Numeric -minimum value of weight to be included in actual. Raise to remove volatility in "actuals" seen in plots}

\item{rebase}{Optional: logical - should the scores be rebased to best match the actuals}
}
\value{
a list of 2 ggplot objects. Each of these plots shows one score split by the other.
}
\description{
plot to inform which of two models is closer to the actuals. This is done by faceting one score by the other and inspecting if the trend is flat
}
\examples{

actual <- rnorm(1000)
incumbent_pred <- actual + (rnorm(1000) * 0.1)
proposed_pred <- actual + rnorm(1000)

plot_2way_comparison(actual=actual,
                     incumbent_pred=incumbent_pred,
                     proposed_pred=proposed_pred)

}
