% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gini.R
\name{plot_gini}
\alias{plot_gini}
\title{plot_gini}
\usage{
plot_gini(
  actual,
  predicted,
  weight = rep(1, length(actual)),
  na.rm = FALSE,
  use_plotly = TRUE
)
}
\arguments{
\item{actual}{Array[Numeric] - Values we are aiming to predict.}

\item{predicted}{Array[Numeric] / DataFrame[Numeric] - Values that we have predicted.}

\item{weight}{Optional: Array[Numeric] - Weighting of predictions. If NULL even weighting is used}

\item{na.rm}{logical. Should missing values be removed?}

\item{use_plotly}{Optional: boolean - If TRUE plotly object is returned else ggplot2 object}
}
\value{
plotly/ggplot object of showing gini curve for all predictions given
}
\description{
Returns a plot showing the gini of one or more models compared to the perfect model and the Null model

Note: Predictions should be annualised (independent of exposure)
}
\examples{

# Input as vector
actual <- rnorm(100, mean=100, sd=10)
weight <- rep(1,100)
predicted <- actual + rnorm(100, mean=0, sd=1)

plot_gini(actual, predicted, weight)

# Input as dataframe

predicted <- data.frame(pred1 = actual + rnorm(100, mean=0, sd=1), pred2 = rnorm(100, mean=0, sd=1))
plot_gini(actual, predicted, weight)

}
\seealso{
\code{\link{metric_gini}}
}
