% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_magic_carpet.R
\name{plot_magic_carpet}
\alias{plot_magic_carpet}
\title{plot_magic_carpet}
\usage{
plot_magic_carpet(
  feature,
  feature_name,
  incumbent_pred,
  proposed_pred,
  weight = rep(1, length(feature)),
  n_bins = 10,
  ratio_max = 1,
  ratio_step = 0.05,
  position = "fill"
)
}
\arguments{
\item{feature}{array[numeric/character/factor] - value of feature}

\item{feature_name}{character - name of feature}

\item{incumbent_pred}{array[numeric] - values of incumbent prediction}

\item{proposed_pred}{array[numeric] - values of proposed prediction}

\item{weight}{array[numeric] - weight of each row}

\item{n_bins}{intergerish - number of buckets to split each (numeric) feature into}

\item{ratio_max}{numeric - max ratio at which to cap the incumbent_pred/proposed_pred ratio}

\item{ratio_step}{numeric - step size to divide ratio bins.}

\item{position}{character - either \code{"fill"} or \code{"stack"}.
If \code{"fill"} all bars are the same hight and extra line is added to show relative population.
If \code{"stack"} bar hight gives population.}
}
\value{

}
\description{
This plot visually compares how predictions change over a factor.
The ratio of two predictions \code{proposed_pred/incumbent_pred} is calculated for each row. Two plots are produced which display
For each level of the feature we find what is the proportion of row with each ratio
For each value of the ratio what is the proportion of row with each value of the feature
The log2 is used to calculate the ratio as this is symmetrical
the values of \code{log2(proposed_pred/incumbent_pred)} which are plotted are -ratio_max, -(n * ratio_step), -((n-1) * ratio_step), ..., -ratio_step, 0, ratio_step, ..., (n-1) * ratio_step, n * ratio_step, ratio_max
}
\examples{
n=100
feature1 <- seq(-10, 10, length.out=n)
feature2 <- rep(c("a", "b", "c"), each=ceiling(n/3))[1:n]
incumbent_pred <- 100 + rnorm(n)
proposed_pred <- 100 + rnorm(n) + seq(-10, 10, length.out=n)

plot_magic_carpet(feature=feature1, feature_name="example feature",
                  incumbent_pred = incumbent_pred,
                  proposed_pred = proposed_pred)

plot_magic_carpet(feature=feature2, feature_name="example feature",
                  incumbent_pred = incumbent_pred,
                  proposed_pred = proposed_pred)

}
