% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PrecisionRecall.R
\name{plot_PrecisionRecall}
\alias{plot_PrecisionRecall}
\title{plot_PrecisionRecall}
\usage{
plot_PrecisionRecall(
  actual,
  predicted,
  weight = rep(1, length(actual)),
  na.rm = FALSE,
  use_plotly = TRUE
)
}
\arguments{
\item{actual}{Array[Numeric] - 0 or 1 - Values we are aiming to predict.}

\item{predicted}{Array[Numeric] / DataFrame[Numeric] - Between 0 and 1 - Values that we have predicted.}

\item{weight}{Optional: Array[Numeric] - Weighting of predictions. If NULL even weighting is used.}

\item{na.rm}{logical. Should missing values be removed?}

\item{use_plotly}{Optional: boolean - If TRUE plotly object is returned else ggplot2 object}
}
\value{
plotly object of showing Precision-Recall curve for all predictions given
}
\description{
Returns a plot showing the Precision-Recall curve of one or more models compared to the Null model

#' Note: Predictions should be annualised (independent of exposure)
}
\examples{

data <- data.frame(x1=runif(100), x2=runif(100), noise=rnorm(100, sd=0.2)) \%>\%
  mutate(target=ifelse(x1 + noise>0.5, 1, 0))

plot_PrecisionRecall(actual=data$target, predicted=data$x1)
plot_PrecisionRecall(actual=data$target, predicted=data[c("x1","x2")])

}
\seealso{
\code{\link{metric_PrecisionRecall}}
}
