% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_numerical_buckets.R
\name{plotting_numerical_buckets}
\alias{plotting_numerical_buckets}
\title{plotting_numerical_buckets}
\usage{
plotting_numerical_buckets(
  var_to_band,
  n_bins = 25,
  weight = rep(1, length(var_to_band)),
  include_outliers = FALSE
)
}
\arguments{
\item{var_to_band}{Vector[numeric] - vector of values to be binned}

\item{n_bins}{numeric - Number of bins to split exposure into}

\item{weight}{numeric - vector weight for observations}

\item{include_outliers}{logical - should outliers be trimmed}
}
\value{
data.frame with columns
bin [numeric]: numeric label of bin
labels [character]: string output from \code{\link{prep_num_bin}} if called
lower [numeric]: Min bound of the bin
upper [numeric]: Max bound of the bin
center [numeric] : Center of the bin - (upper + lower) / 2
width [numeric] : Width of the bin - (upper - lower)
}
\description{
This function takes a numeric vector \code{var_to_band} and groups it into buckets which are useful for plotting.
These groups have approximately equal weight
Outliers are excluded/included using the \code{include_outliers} option.
This function is built on \code{\link{prep_num_bin}}
}
\examples{
plotting_numerical_buckets(var_to_band=runif(100), n_bins=10)
}
