% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_num_bin.R
\name{prep_num_bin}
\alias{prep_num_bin}
\title{prep_num_bin}
\usage{
prep_num_bin(
  var_to_band,
  n_bins = 10,
  weight = rep(1, length(var_to_band)),
  method = "even_weight",
  use_labels = FALSE,
  mean = 0.5,
  sd = 0.3
)
}
\arguments{
\item{var_to_band}{Vector[numeric] - vector of values to be binned. Should be true numeric however can also be categorical with fewer levels than \code{n_bins}}

\item{n_bins}{numeric - Number of bins to split exposure into}

\item{weight}{numeric - vector weight for observations}

\item{method}{string - One of \code{c("even_weight", "gaussian_weight")}.
If \code{"even_weight"} is used the weight (exposure) in each bucket is even
If \code{"gaussian_weight"} is used the weight look like a normal distribution centered on \code{mean} with width \code{sd}. The effect of this is to give thinner tails}

\item{use_labels}{logical - should the bins be numbered (\code{FALSE}) or human readable label such as [1 - 7] (\code{TRUE})}

\item{mean}{numeric - Only used when \code{method}=\code{"gaussian_weight"} - value between 0 - 1 for if the bucket with maximum weight is at min or max prediction. Default is 0.5 (middle bucket)}

\item{sd}{numeric - Only used when \code{method}=\code{"gaussian_weight"} - value between 0.1 - 1 for how thin the distribution tails should be. Low numbers give thin tails}
}
\value{
a list with the following named entries:
bins: Either a numeric vector or a factor (depending on \code{use_labels}) which gives the bin. The vector is of length \code{length(var_to_band)} and has \code{n_bins} unique values
labels: Human readable labels of where binning has occured
cut_values: The numbers the cuts are made at such that they can be reproduced using \code{cut} fucntion
vector with values 1 to \code{n_bins} indicating bin of var_to_band. (1 is low \code{var_to_band} and \code{n_bins} is high \code{var_to_band})
}
\description{
Function which bins numeric vector into bins of equal weight (or a smooth function on weight).
}
\examples{

out1 <- prep_num_bin(var_to_band=1:20, n_bins=5)
out2 <-prep_num_bin(var_to_band=1:20, n_bins=5, method="gaussian_weight")
out3 <-prep_num_bin(var_to_band=1:20, n_bins=5, method="gaussian_weight", use_labels=TRUE)

#To reproduce using cuts
out1$bins
cut(1:20, breaks=out1$cut_values, labels=FALSE)
}
